+++ app/AttributeValue.php
@@ -5,11 +5,13 @@ namespace App;
 use Illuminate\Database\Eloquent\Model;
 use MStaack\LaravelPostgis\Eloquent\PostgisTrait;
 use App\Traits\CommentTrait;
+use Spatie\Activitylog\Traits\LogsActivity;
 
 class AttributeValue extends Model
 {
     use PostgisTrait;
     use CommentTrait;
+    use LogsActivity;
 
     protected $table = 'attribute_values';
     /**
@@ -47,6 +49,11 @@ class AttributeValue extends Model
     protected $postgisFields = [
         'geography_val',
     ];
+
+    protected static $logOnlyDirty = true;
+    protected static $logFillable = true;
+    protected static $logAttributes = ['id'];
+    protected static $ignoreChangedAttributes = ['user_id'];
 
     const patchRules = [
         'certainty' => 'integer|between:0,100',