+++ app/Entity.php
@@ -4,10 +4,14 @@ namespace App;
 
 use App\Traits\CommentTrait;
 use Illuminate\Database\Eloquent\Model;
+use Nicolaslopezj\Searchable\SearchableTrait;
+use Spatie\Activitylog\Traits\LogsActivity;
 
 class Entity extends Model
 {
     use CommentTrait;
+    use SearchableTrait;
+    use LogsActivity;
 
     /**
      * The attributes that are assignable.
@@ -20,6 +24,7 @@ class Entity extends Model
         'name',
         'user_id',
         'geodata_id',
+        'rank',
     ];
 
     protected $appends = [
@@ -30,6 +35,20 @@ class Entity extends Model
     protected $with = [
         'user',
     ];
+
+    protected $searchable = [
+        'columns' => [
+            'entities.name' => 10,
+        ],
+        'joins' => [
+
+        ],
+    ];
+
+    protected static $logOnlyDirty = true;
+    protected static $logFillable = true;
+    protected static $logAttributes = ['id'];
+    protected static $ignoreChangedAttributes = ['user_id'];
 
     const rules = [
         'name'              => 'required|string',