+++ app/ThLanguage.php
@@ -3,9 +3,17 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Spatie\Activitylog\Traits\LogsActivity;
 
 class ThLanguage extends Model
 {
+    use LogsActivity;
+
+    protected static $logOnlyDirty = true;
+    protected static $logFillable = true;
+    protected static $logAttributes = ['id'];
+    protected static $ignoreChangedAttributes = ['user_id'];
+
     protected $table = 'th_language';
     /**
      * The attributes that are assignable.
@@ -17,7 +25,6 @@ class ThLanguage extends Model
         'display_name',
         'short_name',
     ];
-
     public function user() {
         return $this->belongsTo('App\User');
     }