+++ resources/js/app.js
@@ -81,6 +81,8 @@ import {
     faLink,
     faList,
     faLongArrowAltDown,
+    faLongArrowAltLeft,
+    faLongArrowAltRight,
     faLongArrowAltUp,
     faMagic,
     faMapMarkedAlt,
@@ -134,13 +136,16 @@ import {
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserClock,
     faUserCheck,
     faUserCog,
     faUserEdit,
     faUsers,
     faUserTimes,
     faVolumeMute,
-    faVolumeUp
+    faVolumeUp,
+    faDatabase,
+    faIndent
 } from '@fortawesome/free-solid-svg-icons';
 import VModal from 'vue-js-modal';
 import Axios from 'axios';
@@ -168,6 +173,8 @@ import Roles from './components/Roles.vue';
 import Preferences from './components/Preferences.vue';
 import UserPreferences from './components/UserPreferences.vue';
 import UserNotifications from './components/UserNotifications.vue';
+import UserActivity from './components/UserActivity.vue';
+import GlobalActivity from './components/GlobalActivity.vue';
 import UserProfile from './components/UserProfile.vue';
 const DataModel = () => import(/* webpackChunkName: "group-bib" */ './components/DataModel.vue')
 const DataModelDetailView = () => import(/* webpackChunkName: "group-bib" */ './components/DataModelDetailView.vue')
@@ -179,6 +186,9 @@ import VeeValidate from 'vee-validate';
 import Notifications from 'vue-notification';
 import SpacialistPluginSystem from './plugin.js';
 import VueScrollTo from 'vue-scrollto';
+import InfiniteLoading from 'vue-infinite-loading';
+import VueHighlightJS from 'vue-highlightjs';
+import 'highlight.js/styles/atom-one-dark.css';
 
 import { EventBus } from './event-bus.js';
 
@@ -222,6 +232,7 @@ library.add(
     faCopy,
     faCopyright,
     faCubes,
+    faDatabase,
     faDotCircle,
     faDownload,
     faDrawPolygon,
@@ -252,12 +263,15 @@ library.add(
     faFolder,
     faGlobeAfrica,
     faIdBadge,
+    faIndent,
     faInfoCircle,
     faLayerGroup,
     faLightbulb,
     faLink,
     faList,
     faLongArrowAltDown,
+    faLongArrowAltLeft,
+    faLongArrowAltRight,
     faLongArrowAltUp,
     faMagic,
     faMapMarkedAlt,
@@ -313,6 +327,7 @@ library.add(
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserClock,
     faUserCheck,
     faUserCog,
     faUserEdit,
@@ -386,6 +401,8 @@ Vue.use(VeeValidate);
 Vue.use(Notifications);
 Vue.use(DatePicker);
 Vue.use(VueScrollTo);
+Vue.use(InfiniteLoading);
+Vue.use(VueHighlightJS);
 
 const router = new VueRouter({
     scrollBehavior(to, from, savedPosition) {
@@ -541,6 +558,14 @@ const router = new VueRouter({
             }
         },
         {
+            path: '/activity/g',
+            name: 'globalactivity',
+            component: GlobalActivity,
+            meta: {
+                auth: true
+            }
+        },
+        {
             path: '/editor/dm',
             name: 'dme',
             component: DataModel,
@@ -575,6 +600,14 @@ const router = new VueRouter({
             path: '/notifications/:id',
             name: 'notifications',
             component: UserNotifications,
+            meta: {
+                auth: true
+            }
+        },
+        {
+            path: '/activity/u',
+            name: 'useractivity',
+            component: UserActivity,
             meta: {
                 auth: true
             }
@@ -669,6 +702,8 @@ import AttributeSearch from './components/AttributeSearch.vue';
 import CsvTable from './components/CsvTable.vue';
 
 // Reusable Components
+import ActivityLog from './components/ActivityLog.vue';
+import ActivityLogFilter from './components/ActivityLogFilter.vue';
 import UserAvatar from './components/UserAvatar.vue';
 import Attributes from './components/AttributeList.vue';
 import EntityTree from './components/EntityTree.vue';
@@ -693,6 +728,8 @@ Vue.component('entity-type-search', EntityTypeSearch);
 Vue.component('label-search', LabelSearch);
 Vue.component('attribute-search', AttributeSearch);
 Vue.component('csv-table', CsvTable);
+Vue.component('activity-log', ActivityLog);
+Vue.component('activity-log-filter', ActivityLogFilter);
 Vue.component('user-avatar', UserAvatar);
 Vue.component('attributes', Attributes);
 Vue.component('entity-tree', EntityTree);