+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
   - Get a notification when a new comment is posted (and the notified user has already commented on this resource)
 - User profile to allow users to edit their information
   - Support for phone numbers and avatars
+- Activity Log (Logging changes to all models (Entities, Files, ...))
 ### Fixed
 ### Changed
 - Switch from momentjs to dayjs for dates